feat: Add support for setting bandwidth
This feature allows users to specify the bandwidth in the configuration file
or through arguments. The current bandwidth setting is set to a value of
200,000 bits per second as per the previous hardcoded value. In order to
override the default value via arguments, the flag -b or --bandwidth must be
used. The value taken by this argument is an int. In the configuration file,
`bandwidth` can be set to override the existing value and takes a numerical
value represented by bits. The configuration also makes note that the server
will lower the bots bitrate if it exceeds the servers. This is stated in the
mumble documentation https://wiki.mumble.info/wiki/Murmur.ini#bandwidth.