Properly fix version handling with packaging>=22.0
Both the current version and the last startup version can be unparsable.

Hence we now just catch the resulting exception and reset the version
to the new version, which may or may not be parsable, which we'll find
out on the next start up. But either way this prevents crashes from
unparsable versions in the database.